Update patch set 2
Patch Set 2:

I'm not following also :)

I mentioned this needs a redirect (current patch leaves a 404 at Special:History for existing urls all over the internet, including on-wiki in discussions and edit comments). To which you responded yes, but you need to convince a designer and PM first.

This was surprising given a HTTP redirect doesn't have a UI, and not breaking URLs doesn't seem like a product decision (that is, keeping a redirect is part of the cost of implementing this change, it's not really something a PM even needs to be aware of per se).

But yes, keeping a simple redirect is indeed all that's missing, and all that's needed :)

Anyway, in thinking about this I wrote a basic one that seems to work locally. Feel free to use as starting point. https://phabricator.wikimedia.org/P8361

Also fixed up something that wasn't working right in mediawiki/core while at it.

Patch-set: 2
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>