Add link to current article from Special:MobileDiff page
Adds a link to the current version of the article from the
Special:MobileDiff page and changes CSS so that links look
like links instead of headers.

Converts the PHP output to utilize a mustache template instead
of the HTML:Element class.

Replacing the `mobile-frontend-diffview-comma` message in favour
of using the `comma-separator` message.

Bug: T200969
Change-Id: If8cec485ba22eb96a9f9022308f1535ef957e231