Update patch set 2
Patch Set 2:

Thanks, that makes sense.

The confusion was between having a redirect (when replacing the mobile version) and the decision itself to replace it at all.

My questioning of manager input was about the former, not the latter. Of course, whether to replace it at all and whether it's ready, indeed makes a lot of sense to need such input no. No question there :)

Patch-set: 2
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>