Disable broken Ruby browser tests
Any Ruby browser test with a login step no longer works. This
adds noise to the ones that do, so should no longer be run in the
daily job.

This does however mean these browser tests are never run. They
will be restored slowly but surely by porting them to Node.js
and will be prioritised in that porting first.

Bug: T219920
Change-Id: Ie969256cba2d45c27cb48064f4e880b6cac6bebf