Store Page.js headings in instance variable
On pages with a ton of nodes like
https://en.m.wikipedia.org/wiki/List_of_Nvidia_graphics_processing_units,
the pageIssues code would take over 2.5 seconds to execute and was
blocking the main thread. The jquery .find call in Page.js was
particularly expensive.

By the headings in an instance variable, we can limit the
blocking of this from around 2.5 seconds to 50ms!

Note: I do not know how the page issues code works and if the headings
can be cached! Please double check my work.

Bug: T220751
Change-Id: If4f0bca79c4531c7be20ae17ad0ebe0b15cc75d0