Update patch set 2
Patch Set 2: Code-Review+2

Sometimes an unused variable can still have meaning. For example, if it holds a scoped destructor callback. Or if the value assigned to it was created by a function that had side-effects.

For example:

 preparePlatform();
 $success = launchRocket();
 catchRocket();

$success is unused, but removing the line means the rocket is no longer launched.

In this case, the method used has no significant side-effects and is indeed fully unused. You probably know this already, but I thought it was a funny example.

Good day!

Patch-set: 2
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2