Update patch set 13
Patch Set 13:

This is too big commit for me to give it a good review.

It:

- removes Ruby Cucumber+Selenium tests
- adds Node.js Cucumber+Selenium tests
- adds Node.js Mocha+Selenium tests

Each of those could be a separate commit.

Also, I disagree on the architecture of Mocha+Selenium tests.

Another commit depends on this one:

https://gerrit.wikimedia.org/r/c/integration/config/+/502356

I would highly recommend splitting this commit into smaller ones. The important thing is to unblock patch 502356. For that only a small commit containing selenium-test npm script and one Mocha+Selenium test is needed.

If that is not an option, I'm fine with this commit being merged. It does contain selenium-test npm script and a Mocha+Selenium, so it's unblocking the other commit.

Patch-set: 13
Reviewer: Zfilipin <487@e9e9afe9-4712-486d-8885-f54b72dd1951>