✅ Selenium tests: Adding switch_views.js
switch_views.js tests the button that toggles the mobile/desktop view.

- The toggle elements have been added to the article_page.js
- the cookie handling of minerva_page.js has been improved to set an
  expiry time (1 hour).
- The page-objects have been used to write the test instead of the
  cucumber.js transition step in order to make the browser state more explicit.
- Documentation on how to point the tests at the beta cluster has been added.

Additional changes:
When wgMFDefaultSkinClass is undefined or points to a skin that has not
been installed use the DefaultSkin (which is definitely available). This makes
our browser tests run against Vector skin in mobile mode. The fallback skin
has no footer so without this change, the browser tests fail (no desktop link
link). This has no impact on production as wgMFDefaultSkinClass is setup correctly.

Bug: T220659
Change-Id: I884ecc4c4206da953c895fa1b17f63382d9aa537