💚 Run user_page test on non-existent user page to fix QA build
The user_page.js test assumed that the user running the test does not
have a user page. These tests are running against the wikipedia beta
cluster where this assumption is not true. Specifically, they are being
run as "Selenium user", which does have a user page.

The test passes on any user page, so we can use a user account that
doesn't exist.

Bug: T220659
Change-Id: I2ddd529b24a19fbb4a61ccaf9679c269b10be9d1