Update patch set 2
Patch Set 2:

> The rule responsible for this behavior is here: https://phabricator.wikimedia.org/diffusion/SMIN/browse/master/resources/skins.minerva.mainMenu.styles/mainmenu.less$182
Are we talking about the same bug?
I've verified again locally by loading http://localhost:8888/w/index.php?title=Return_to_Zork#/notifications and zooming out so the resolution is suitably  large. The problem as I perceive it is the rule  @ screen and (min-width: 1000px) which sets     max-width: 993.3px;

The left rule you point to is what makes this overlay behave as a drawer. If removed it takes up the full screen.

My interpretation of the bug is we want to remove the gutter on the RIGHT of the notificationd drawer (see https://phabricator.wikimedia.org/T219320#5136474)

Patch-set: 2
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>