Swap Hogan for Mustache
This replaces Hogan template rendering with the better supported
mustache rendering that is defined in core, and seeing more github
activity.

The Hogan module is deprecated and redirects to the Mustache compiler
to allow us time to update usages inside Minerva.

Importantly, there appears to be no difference in JS before/after this change
(post gzip) and the workflows for all features continue to be usable.

Bug: T220620
Change-Id: If00010e092e3766f9812f9a69ce0dfe22298689a