Update patch set 1
Patch Set 1: Code-Review-1

Given the jenkins -1 and the fact that patches in core are not experiencing problems with https://gerrit.wikimedia.org/r/502820 , I think this test is a red herring to what's actually the problem here.


I'm seeing a token error, in the console output on https://gerrit.wikimedia.org/r/#/c/mediawiki/core/+/506429/ so it's possible the Selenium tests in core are using the API to logout but passing the wrong token and will need to be updated.

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1