💚Do not rely on ArticlePage.setMobileMode()
This seems to be problematic when a mobile domain is used e.g.
beta cluster. Instead of using it, switch to mobile site and then
back using the links at the bottom of the page, verifying you end
up back on desktop.

The two tests are merged to avoid having to do setup and teardown
(the first test left the user on the mobile page, when the second
test begins in desktop mode.)

Bug: T220659
Change-Id: I143ae653a7f7f8781cfb268ce46ffe8362d58c84