Remove the lazy load references beta feature
The lazy load references feature was added alongside the lazy
image in August 2016. Since then it has sat in beta for a period of 3
years without any work.

The main blockers for promoting it were around the slowness of the
references API that we created.

Reasons to remove it:
* 3 years and no activity
* It uses the mobile view API which we'd like to deprecate
* It's buggy and significant work is needed to complete it (see subtasks of goal
task)
* We're supporting it with a lot of code
* It complicates our refactor project

Bug: T222373
Depends-On: I2dcf9b0cbfe2f46eb6763e0add4ed892d1fbecd9
Change-Id: I373f4ca8115d0db73b153848872b97b9423248d7