Rename EditorOverlay to SourceEditorOverlay
There are three classes that are responsible for two editor overlays:
* EditorOverlayBase
  * EditorOverlay (now SourceEditorOverlay)
  * VisualEditorOverlay

I think SourceEditorOverlay is a more intuitive name than EditorOverlay.
It reflects the inheritance better (and results in alphabetical
ordering matching it), and makes it easier to tell at a glance whether
some piece of code works with both editors or only the source one.

There is some bikeshedding to be had on whether it should be called
SourceEditorOverlay or WikitextEditorOverlay. I picked the former
because we've already usually called it the "source editor" when it
was necessary to distinguish the two (e.g. in the 'preferredEditor'
preference).

Also fix some comments that erroneously referred to EditorOverlay but
they really meant the base class or both of the child classes.

Change-Id: Ibb3eeac483ac7736a190dbb1efb9dc3c1cde60b8