ContentProviderFactoryTest: Remove unused vars in test suits
$provider was defined but never used and in fact, these particular
cases just needed to throw an exception that will match exception type
set just above them. We don't need to keep this in a variable. Once the
exception is thrown, the assertion is performed.

The code that introduced this is 52a8dc9, written by me so I can confirm
these vars were not used. Also, I put the setException() call just above
the code that will throw the exception to reduce cognitive load and follow
best practice.

Change-Id: If639aace933a088cf4591d2d82bd22c9d6763493