Replace use of `MobileContext::singleton()` with MC service
This change fully moves the use of `MobileContext::singleton()` into
the use of MobileContext service in preparation of breaking the MC
class into smaller services.

Also, does some cleanup in `ServiceWirings.php` related to namespaces
and imports. In addition, soft deprecated the `getMFConfig()` and the
`singleton()` method with the replacement of its use by services.

Bug: T218244
Change-Id: Iee53988dcd46f5b20fcd33ffbfe4177cf99c6893