Update patch set 1
Patch Set 1:

This patch probably does no harm, but it also does not make anything better, I'm afraid.

See, marking a method as "static" means it *can* be called with self::…, but it does not need to.

The fact this method is not really used
https://codesearch.wmflabs.org/search/?q=escapeWhiteSpace
is more a hint that the "static" keyword should be removed, and the method made "protected".

Patch-set: 1
CC: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>