Update patch set 5
Patch Set 5:

(2 comments)

The patch looks good and I believe it can be merged as it is. But the cleanup started here is not fully finished, it seems. See inline comments. It's certainly possible to continue working on this code in later patches, so this shouldn't block this one.

Patch-set: 5
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>