Update patch set 1
Patch Set 1:

Most probably escapeWhiteSpace() was declared static as it doesn't use any of TableDiffFormatter properties. It's a helper method that allows you to quickly escape strings.

My first thought was to keep the "static" keyword, but then I noticed that `blockHeader()` also doesn't use $this, and it is not static. 

I agree with Thiemo, most probably it would be the best to remove that single `static` declaration, and make it a method of TableDiffFormatter. Honestly, I think it can be protected (as only InlineDiffFormatter and TableDiffFormatter use those).

Patch-set: 1
Reviewer: Pmiazga <3945@e9e9afe9-4712-486d-8885-f54b72dd1951>