Update patch set 1
Patch Set 1: Code-Review-1

Where is this necessary?
The only inputs I see in MobileFrontend are using OOUI or mediawiki UI (or the search input in the header which uses neither) which define padding .

Can you give an example of a page where button and input do not have padding and this is problematic? Using the default browser inpu/button padding seems preferable to me in this situation.

-1 until I have more clarity about the test cases you are using.

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1