Mobile VE: Show fake toolbar while loading
The new toolbar with a spinner replaces the old loading progress bar.

The real VisualEditor toolbar depends on basically all of OOUI, and
we don't want to load it immediately, so instead we create a plain
HTML replica of it, with just enough copy-pasted CSS (and a single
OOUI icon, 'close') to make it look like the real thing.

In the future, if we arrive at a common design for MF overlay headers
and the VE toolbar, maybe we can remove all of this and instead just
add a header to the loading overlay.

Bug: T217784
Change-Id: If21aa0ea619ec2500ce5fca6fe81eb27f26bb047