🐛 Bug: Add border back to edit overlay summary textarea
It was missing because of an overly broad css rule with panels &
input/textarea elements. This moves the rule to only apply in
TalkSectionAddOverlay when the panel contains either an input element or
textarea and nothing else (which is the only case where this is used).

A new ticket will be created to consolidate these form styles in
overlays (I tried to do that in
I30aa36a0cb89988d1bad8256e5d3f737380bc0a7) but it did not reach a
consensus and deserves more discussion.

Bug: T221624
Change-Id: I572350a201fbc859bb41a4ef2ce5d8875e164ba5