Update patch set 5
Patch Set 5: Code-Review+2

I agree with the comments above. It might not seem like the neatest solution, but setting  properties on the prototype is the most efficient way of going about this. As I mentioned to Nick in chat, we could in theory also pass properties as arguments to super(), but that would require us refactoring View to make it accept those arguments.

Patch-set: 5
Reviewer: Jdrewniak <2950@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2