EditorOverlayBase: Fix browser window close warning after switching
Due to the order of operations, when we were switching from
SourceEditorOverlay to VisualEditorOverlay (or vice versa), the second
editor's warning was registered first by its constructor and then
immediately released by the first editor's onBeforeExit().

Move the code from constructor to show() so that things happen in on
order that works.

Bug: T222315
Change-Id: I3c7a07a1a51ba856f9d96765fb64708e4b6f93e3