Update patch set 3
Patch Set 3:

I think the config options wgMFUsePreferredEditor and wgMFDefaultEditor are redundant.

If we want to forever support setting a default editor mode other than on desktop, then we should have just one config option, with the values of 'source', 'visual', 'preferences', and 'abtest' (or equivalent).

If we want to only introduce this temporarily to run our own A/B test, then instead of wgMFDefaultEditor we should just have a boolean to enable the A/B test.

I note that we can still remove or rename wgMFUsePreferredEditor if we want, since it wasn't included in a release yet (but will be soon when 1.33 goes out).

Patch-set: 3
Reviewer: Bartosz Dziewoński <417@e9e9afe9-4712-486d-8885-f54b72dd1951>