Update patch set 3
Patch Set 3:

I am not sure if it's helpful to split things off to a separate file (also, we should clarify what we wanted to be split off – I understood Niedzielski's comment to be just about the newly added code, as opposed to all editor-related code).

But if we do split it, then we should do it in this commit (or a preceding one) rather than a following one, so that we can cleanly remove the A/B test code after the test concludes by reverting this commit.

Patch-set: 3
Reviewer: Bartosz Dziewoński <417@e9e9afe9-4712-486d-8885-f54b72dd1951>