Allow AMC users to see core Special:Contributions page
If AMC is enabled, users will see the Special:Contributions page from
core. If disabled, users will see the SpecialMobileContributions page.

Note: The DateInputWidget will not initialize because the
`mediawiki.special.contributions` module's target does not explicitly
include `mobile` [1]. We can enable this, but it will add 80KB to the JS
payload!

[1] https://github.com/wikimedia/mediawiki/blob/ba7f053c556626339cd8bc56f61b0177115e4b5d/resources/Resources.php#L2124-L2129

Bug: T225124
Change-Id: I3df0b56b1e965138c36bbe317de5187b135e19b1