Remove Page.js inheritance of View.js
This commit attempts to be very incremental with the removal of the
inheritance by getting rid of the defaults key on the prototype (and
moving that into the constructor), moving the preRender logic into the
constructor, adding a $el property (similar to View) so that some of the
HTML parsing methods still work.

⚡ Disclaimer: When moving the `defaults` code I realized that there are
a set of instace variables on Page.js and a set of properties on
this.options and those could in fact be different for the same key (e.g.
this.thumbnail vs. this.options.thumbnail, this.options.sections vs.
this.sections) which I have found very confusing. I'm not sure why we do
this but am looking how to address this in a follow up patch.

Bug: T193077
Change-Id: I46731e416f9edfb9b74c97435dc10fd1297f690e