Drop resource-modules
resource-modules protects against importing undeclared dependencies
defined using M.define and used via M.require as well as missing
templates and messages from ResourceLoader modules. Only the message
portion of this is useful to this project, but the value of this is
not worth the investment.

Given there is problems with this in the node 10 port, we have decided
to remove these and leave the resource-modules as unofficially unmantained
for the time being.

Bug: T224997
Change-Id: I4dab9455bde7b9b24bf62dc0a975e489509bb55d