Refactor MF to use PageHTMLParser / clean up Page.js usage
In I2f32dbcc4ebaa4bebb297cda1ecce3457922b343, Minerva was refactored to
make use of the refactored Page.js class and the new PageHTMLParser.js
class. This commit refactors MF to use PageHTMLParser as well as cleans
up prior Page.js logic.

Notable changes:

* Toggler.js was refactored to remove the setting of Page instance data
on each $heading. I'm not sure what the original purpose of this was,
but I think this is unnecessary since Toggler.js already has the page
instance injected into its constructor function. In order to support
this removal, the method signatures of this class were changed to allow
the passing of Page.js instance. I debated about making this an instance
variable instead (but chose not to because the Toggler class seems to
function more as a script with side-effects rather than a class that
clients can call). LMK if you feel it is better as an instance variable
and I will happily make the change though bc I don't feel strongly about
it.

* Closely related to ^^, the `section-toggled` event in Toggler was
changed to not pass the current page instance. Skin.js seems to be the
only client of this event that needs the page instance yet it already
has one injected into its constructor which it then passes to
lazyReferencesLoader so this was eliminated.

* lazyReferencesLoader previously used the page instance from its data
`arg` as well as the page instance from its `page` arg. I didn't
understand this so I instead made it only use its passed in `page` arg
when it need a page instance.

* Perhaps controversially, the ReferenceGateway's `getReference` method
signature was changed to accept a Page instance AND a PageHTMLParser
instance. ReferencesHtmlScraperGateway only needs to use PageHTMLParser
while ReferencesMobileViewGateway only needs to use Page. Still,
minerva's references.js script is setup rn to allow these classes to be
interchangeable and I found this to be the simplest solution rather than
do more extensive refactoring.

Bug: T193077
Depends-On: I2f32dbcc4ebaa4bebb297cda1ecce3457922b343
Change-Id: I6e858bbe73f83166476b5b2c0fdac1cca7404246