MoveLeadParagraphTransform: Simplify isNotEmptyNode()
We don't need to call isNotEmptyNode() recursively for all descendants.
$node->nodeValue already includes the text of all descendant nodes.

While we're here, change the ->nodeValue to ->textContent. They are
exactly the same, but the latter is a more intuitive name, already
used elsewhere in this file, and consistent with the JavaScript DOM
implementation found in web browsers.

The remaining code can be simplified to a single line.

Ref:

* https://www.php.net/manual/en/class.domnode.php#domnode.props.textcontent
  > The text content of this node and its descendants.

* https://www.php.net/manual/en/class.domnode.php#domnode.props.nodevalue
  > The value of this node, depending on its type. Contrary to the W3C
  > specification, the node value of DOMElement nodes is equal to
  > DOMNode::textContent instead of NULL.

Change-Id: Icd2e247b91a0d8331aca6f7eb99090aaca229c5f