mobile.editor.overlay: Scroll to the lead paragraph on load
Due to the lead paragraph transformation in read mode, users may be
confused to see the infobox first when opening the editor for the
lead section. (And to a smaller degree, invisible templates and slugs
caused by the presence of hatnote templates cause similar problems.)

To help users stay oriented, do our best to keep the lead paragraph
in the same positon on the screen when opening the editor.

Reimplemented MoveLeadParagraphTransform::identifyLeadParagraph()
in JavaScript, so that we can run it on the edit mode HTML. Added
some doc comments so that hopefully the two versions stay in sync.

Bug: T225936
Change-Id: I00637c4801644859879c42fee17d499033bdb3a9