Super-light refactor of SearchOverlay
The SearchOverlay is a bit of a monster. This begins the process
of pulling out SearchResultsView into its own component.

This timeboxed (1hr) refactor starts separating the concerns of
the display of the search results from the search itself and I think
although incomplete is a step in the right direction.

This is the first ES6 class in startup so hence the increase in
mobile.startup bundlesize. This will not increase for further adoption
of ES6 classes

Bug: T220849
Change-Id: Ia5fa95b5bb82e6931dbb9cb3387345eeb74ead72