Clicking links triggers onBeforeExit
The onBeforeExit function is now taking a cancel argument.
This is going to be needed if we go in a Redux direction, so I
think this makes sense.

Any links clicked will now trigger the function, hitting ev.preventDefault
if the user says they are not willing to leave the page

Bug: T229701
Change-Id: I2a5de0652d6dabc13b02aac4a65995c05662547f