Overlay: Make .hide() + .show() in rapid succession work
If two routes are mapped to the same overlay (meaning, the same object)
in OverlayManager, then switching between these routes will cause that
overlay to be hidden and immediately shown again. This is fine in
principle, but it breaks in practice: calling .hide() then .show() on
the same overlay in rapid succession results in the overlay being
hidden.

This is caused by the deferred .detach() call scheduled by .hide().
To fix this, have .show() cancel the deferred detach if it hasn't
already happened.

Change-Id: Idac6553789ff224e3584f730f9b657f295e2052f