[fix] [menu] [UI] [notifications] notification drawer closing animation
When closing the notification Overlay, it was first detached from the
DOM and then animated away. This patch waits for the animation to
complete and then detaches the Overlay.

As part of this patch, calls to Overlay.onBeforeExit() in Overlay and
OverlayManager were updated to use the Overlay as context. This
shouldn't break any existing usages and seems more intuitive.

Bug: T206354
Change-Id: Idaeff70ac87cfde32878a1f22657f0567fbe7545