🐛 Fix SpecialMobileOptions.php support of `returnto` redirects/single form submissions
This is needed in order for the BetaOptInPanel, etc to submit a form and
be redirected back to the page the user submitted the form from. It
looks like at one point in time the BetaOptInPanel used the redirect
functionality but somewhere along the line a regression happened.

Additional changes were added so that forms updating a single field (e.g
BetaOptInPanel) would not clobber other user options.

Both of the functionalities will also be needed in
Iba1e256cdac697e49113a49c0e52dbe5504465e0 with the AMC outreach
campaign.

Bug: T221395
Bug: T226068
Change-Id: I0fddbd57558a127c06848b2546b776cd94ac884f