Add Amc Outreach Drawer Feature
This patch adds the foundation and onload action needed for the AMC
Outreach Drawer.

To see this feature in action make the following edits to your
LocalSettings.php:

1) `$wgMFAmcOutreach = true;`
2) `$wgMFAmcOutreachMinEditCount = 0;` (or keep it the default 100 edits if
you already have made that many edits)

Notable changes:

* The drawer itself is lazy loaded (mobile.amcOutreachDrawer module)
while the business logic associated with whether it is eligible to be
shown, etc is kept in mobile startup (mobile.startup/amcOutreach). The
interface for clients becomes a lot easier to use if the business logic
is not lazy loaded.

* Changed mobile.special.mobileoptions.scripts to support onToggle
callbacks as the AMC toggle needs to save outreach state when it is
toggled off. [1]

[1] https://phabricator.wikimedia.org/T226068#5374080

Bug: T226068
Change-Id: Iba1e256cdac697e49113a49c0e52dbe5504465e0