mobile.init/editor: Fix the scrolling animation after Minerva changes
This line of code is meant to get the position of the clicked heading,
relative to viewport.

As a part of this, it was using the 'offsetTop' property to get the
position of the heading relative to the document, but the property
is actually the position relative to the "offset parent" node (the
closest positioned containing element) rather than the document!
It worked fine by accident because previously the offset parent was
the root node, but in change 83f90dacad5869570bd4a16ff7dd38cca2d92cfb
`position: relative;` was added to `.content`, making it the new
offset parent.

What we really want is to use getBoundingClientRect(), which directly
returns the position relative to viewport, saving us the extra
calculation with the scroll position.

Bug: T230237
Change-Id: I67ef10a9ca50a34fafc5629fd1574b77228e4daa