mobile.init/editor: Simplify some loading code
Before b9b9b8f33fc45d2fcd0763fab9af68d980a2740f, we replaced the
loading overlay with the editor overlay in the middle of the loading
process, which required us to handle the user aborting the loading in
two different places, depending on which overlay was shown.

After that change, we replace the overlay only after everything is
loaded, so we only have to handle aborting on the loading overlay.
This should have been changed in that patch.

Change-Id: Ia830428e89bf69a55f2742b9d08daa4e5719f295