mobile.editor.overlay: Add a consistent way to wait for editor data to load
Previously you would need to do it differently for SourceEditorOverlay
and VisualEditorOverlay, and differently when loading "from scratch"
and switching editors.

The requests to load data are now started immediately when constructing
the overlay rather than when showing it. This makes no difference in
practice because we do both at the same time.

Change-Id: I80f4a290144abb29f8faef14464da86566567701