mobile.editor.overlay: Use consistent transition when switching editors
Previously switching editors would result in the current editor instantly
closing itself and opening the other editor, which displayed some kind
of ugly loading spinner on a completely blank surface.

We want to create an illusion of staying within a single editor with two
modes, and avoid distracting the user from the content they're editing.

To do this, implement a transition similar to what we're using for loading
the editor (in mobile.init). The toolbar is replaced by a fake toolbar,
the editing area fades out, and once the new editor is ready, it appears
seamlessly.

This also simplifies the code in both overlays, as they don't have to
support a mode where the overlay is shown but the content of the editor
is not yet loaded. They are just never shown until they're ready.

* Remove code and styles for the old "loading mode" in both overlays,
  add styles for the new "switching mode".
* Tweak fake toolbar styles so that it works in the editor overlays.

Bug: T228159
Change-Id: I2f2495ab6c10116a6660f4361e49272cb95b988a