mobile.editor.overlay: Avoid editor flickering when the user is blocked
When the user is blocked, we previously opened the editor, then
immediately closed it and displayed an error message. That looks
silly, so don't do it.

This cleans up duplication of the block handling code between the two
editors.

This also fixes a bug where due to a race condition the editor might
stay open, that I caused in I80f4a290144abb29f8faef14464da86566567701.

Change-Id: I326b50ec16e8b74dba6e9669ba479daebef7636d