Fix no-restricted-properties/syntax inheritance
The configuration for "no-restricted-properties" that we set in
.eslintshared.json was previously overriding configuration from
wikimedia/eslint-config-wikimedia, which uses this property to
disallow methods from ES6 and later versions when we target browsers
that only support ES5.

This allowed methods like Object.values to be used, breaking compat
with IE 11. Example: a43ac3237ce4cf5a14c89508ecc601005170c67b.

Use some custom code to merge the arrays, similar to the commit
https://github.com/wikimedia/eslint-config-wikimedia/commit/63a5088c
(and using the 'merge' function defined there).

Change-Id: I7619ce860debdd46d9c87167a2a6bfa7bd5fd963