Make promoCampaign.js API make use of mw.storage return values
Piotr correctly pointed out in super-happy-dev-time that mwStorage
operations can fail (e.g. storage is unavailable, full, etc).
Furthermore, the mw.storage operations will tell the client when an
operation has failed. This commit changes the API to make use of this
by:

1) Return the mw storage operation result from `makeActionIneligible`
method. This actually doesn't have a use case right now but is probably
a good idea for the API to return. amcOutreachDrawer calls this method
before showing a toast message telling the user that they can enable amc
in their settings, but we would probably want the toast message to show
even if the save operation fails.

2) Change `isActionEligible` function to only return `true` if the
mw.storage.get returns `null`. This distinction can be important because if
the get operation fails for any reason (storage is unavailable), we
wouldn't want to show the drawer.

Note: `makeAllActionsIneligible` was delibriately kept as is since
multiple save operations happen in it, and the return value doesn't make
as much sense there. Ideally, and in the future, the save operation
should probably be made atomic by using mw.storage.setObject to set an
object with multiple action keys,  instead of using mw.storage.set and
having a separate storage key for each action.

Change-Id: I750cc45040e4e075ab23cf795a5dd038d173fb77