Separate text element from mw-ui-icon-before
It should be possible when using the mw-ui-icon-before class to change
the font-size of the label but not the font-size of the icon - for
instance you might want a normal size icon next to small text (e.g.
as in the case of the MainMenu)

This is demonstrated with Special:MobileDiff page where we make the
font-size only touch the label.

This change is backwards compatible and will make standardising
icons easier.

PLease test: Special:MobileDiff and all changelist (AMC disabled)
pages on mobile when verifying this change.

Bug: T229440
Change-Id: I3f803ec4c9068b30aa93b803391aa4d65d8310ff