Update MobileFrontend icons to WikimediaUI theme equivalents
* Repurpose mobile.init.icons as mobile.ooui.icons and collapse
mobile.startup.images.variants into it
* Update fakeToolbar to use the icon helpers rather than hardcode classes
* Rename all overlay-close to 'close' (no caching implications - JS only)
* Rename all 'arrow' icons to 'collapse' to reflect OOUI icon pack name
(no caching implications)
* Rename 'back' to 'arrowPrevious' (no caching implications)
* No need to rename 'alert'
* diff icons are now a single icon with 3 variants and rotation (no
caching implications)

Additional changes:
There is no icon in Wikimedia UI for search in pages or toc.
The toc icon is only used in Minerva so has been moved there in the Depends-On
and thus can be removed from this repo.
The icon for search within pages is updated here.

Bug: T231613
Depends-On: I86a4bdac83e1d6caef025ad025962127d74556dd
Change-Id: Ibbc706146710a9e31a72b3c2cd4e247d7a227488