Dev: Include all files in code coverage reports
Loosely inspired by https://github.com/istanbuljs/nyc/issues/594
a new test is added that imports all testable files and
checks that they can be imported. This ensures our code is always
loaded even if it is not tested, that our code's side effects is
well documented and that our code coverage reports are accurate.

It seems the all=true option does not appear to be compatible with
ES6 code and QUnit

Change-Id: I38b274a878d3dfd0e2fa7c337411074be3275fe9