mobile.init/editor: Fix handling loading errors other than being blocked
Turns out there are code paths where the rejection value from mw.Api
"bubbles up" all the way to where we tried to handle the user being
blocked. The value is a string (error code) rather than a Drawer, so
calling .show() was causing exceptions. Display a generic error
message in this case.

(Also remove error handling for this case in SourceEditorOverlay,
which is not doing anything because the overlay is never shown.)

Change-Id: Ie24878d62a99b4c23b5a833ce45f4b359d680a34