Restrict the MobileFormatter to certain pages
Based on the analysis inside:
https://phabricator.wikimedia.org/T204606#5484240

I have chosen some conservative values that match for all the
examples given. The values are configurable so we can tweak
them as necessary.

The config value MFMobileFormatterHeadings (which is not
altered in production) is merged into this new configuration
option.

A follow up will merge more configuration for the mobileformatter
into the new MFMobileFormatterOptions

Bug: T232690
Change-Id: I877473bbeb8410f4959553076a78f7c51e4b473e