Avoid expensive WebRequest::getText() in early init code
The getText() method uses getGPCVal() internally which does
various expensive locale-aware Unicode/UTF normalisation. This
is not neccecary for internal parameters that are not expected
to be entered by users directly or where the expected value does
not contain characters subject to such normalisation.

This is motivated by the proposed next step for T213242 which
will call MobileContext::shouldDisplayMobileView() significantly
earlier than it is called today, and thus avoid triggering various
expensive and lazy-loaded code in that earlier path.

Bug: T213242
Change-Id: I528deb97b1db39ce471bec71d73dc132b56e0a65