Move target=mobile override to 'BeforeInitialize' hook
Setting it from the BeforePageDisplay hook means that for most of
the web requests's life time, OutputPage::mTarget is set
incorrectly. Even even during the BeforePageDisplay hook, some
other extensions using the same hook will also see it incorrectly
because MobileFrontend's callback isn't the first to run.

Set it from BeforeInitialize instead to fix this issue.
More specifically, it means the Gadgets extension can safely
decide from within its BeforePageDisplay hook whether to load
a gadget or not and thus not get a warning from OutputPage about
queuing a desktop-only module on mobile or vice versa.

This in turn is blocking the enablement of that warning in
production, which in turn is blocking the discovery of other
(actual) dependencies on the Target system, which in turn blocks
the removal of the Target system (T127268).

For the rare case of Blacklisted pages (at WMF, only
used on foundationwiki for one legacy special page), we still
set it late. This means Gadgets may continue to emit warnings
there of the kind they emit there today.

Bug: T213242
Change-Id: Ibe693bb6a84bdf56fd22d1ea1f0b798094b91c62