Load editor EventLogging code earlier
Previously, code that records EventLogging data for events we track
in the editor (schemaEditAttemptStep.js) was loaded together with the
editor code. As a result, events related to initialization of the
editor were not recorded if the editor did not load (T232237#5501940).

schemaEditAttemptStep.js and schemaVisualEditorFeatureUse.js are now
moved into 'mobile.init'. This way will allow us to record editor
initialization. Hopefully the added code doesn't have much effect on
performance.

Also move schemaMobileWebSearch.js to the same module (it was
previously part of 'mobile.startup', which is apparently undesirable).

Move the dependency on 'ext.eventLogging' to a single place. I'm
leaving the silly function wrappers so that Git's rename detection
doesn't go crazy due to whitespace changes, they can be removed later.

Bug: T232237
Change-Id: I72799083e31bc6dad5d6e5b2a2437ea5a7300497