Dev: Drop transparent-shield from MobileFrontend Skin.js
Drop has-drawer and drawer-enabled classes in favor of navigation-enabled
and remove the insertion of a transparent-shield inside Skin.js

This DRYs up some code and relies on the navigation-enabled class in
Minerva.

In future all the code inside the Drawer codebases should be moved to Minerva
but that's a patch for another time.

Change-Id: Id07ae4e66e9fc97da86da57642fccbe7b8a7597a