Reorder package.json commands
Reorder npm test commands based on usefulness and speed.

linting JS is more helpful and quicker than building assets
as any issues with linting that need fixing will require changes to
assets, so do that first and report on those problems first.

Similarly, there is no point in building and checking assets until
the tests are passing.

SVG checking is quick and easy to rectify so that is lso pushed up the
queue.

Change-Id: Iee6d2bee49065a2ba9864f2a6389e989eee089c7